Rename logging.py to activity_log.py and update imports
Renamed src/logging.py to src/activity_log.py and updated all relevant import statements across multiple modules to reflect the new filename. This improves clarity and consistency in naming related to activity logging functionality.
