+++ src/scooters.py
@@ -7,7 +7,7 @@ from validation import (
     validate_location,
 )
 from auth import get_current_user, check_permission
-from logging import log_activity
+from activity_log import log_activity
 
 def add_scooter(serial_number, scooter_type, battery_level, status, location):
     """
@@ -526,7 +526,7 @@ if __name__ == "__main__":
 
     # Show logs
     print("\n--- Activity Logs ---")
-    from logging import get_all_logs, display_logs
+    from activity_log import get_all_logs, display_logs
 
     logs = get_all_logs()
     display_logs(logs[-10:])
